define WANN.h
* because I might want to use the WANN class nn.cpp later
